+++ utilities/compacted_db/compacted_db_impl.h
@@ -56,7 +56,7 @@ class CompactedDBImpl : public DBImpl {
   using DBImpl::CompactRange;
   virtual Status CompactRange(ColumnFamilyHandle* column_family,
                               const Slice* begin, const Slice* end,
-                              bool reduce_level = false, int target_level = -1,
+                              bool change_level = false, int target_level = -1,
                               uint32_t target_path_id = 0) override {
     return Status::NotSupported("Not supported in compacted db mode.");
   }